Allow to open projects from the game dashboard
- [x] Display "Open project" button (with arrow if multiples)
- [x] Display cloud/local projects in game dashboard for a game
  - [/] Ensure we always save the game id
  - [x] Adapt the wording if nothing is found.
