+++ newIDE/app/src/MainFrame/EditorContainers/HomePage/BuildSection/index.js
@@ -2,8 +2,6 @@
 import * as React from 'react';
 import { type I18n as I18nType } from '@lingui/core';
 import { Trans, t } from '@lingui/macro';
-import List from '@material-ui/core/List';
-import ListItem from '@material-ui/core/ListItem';
 
 import Text from '../../../../UI/Text';
 import TextButton from '../../../../UI/TextButton';
@@ -20,20 +18,16 @@ import {
   type FileMetadata,
   type StorageProvider,
 } from '../../../../ProjectsStorage';
-import PreferencesContext from '../../../Preferences/PreferencesContext';
 import AuthenticatedUserContext from '../../../../Profile/AuthenticatedUserContext';
 import SectionContainer, { SectionRow } from '../SectionContainer';
 import {
   checkIfHasTooManyCloudProjects,
   MaxProjectCountAlertMessage,
-} from './MaxProjectCountAlertMessage';
+} from '../CreateSection/MaxProjectCountAlertMessage';
 import { ExampleStoreContext } from '../../../../AssetStore/ExampleStore/ExampleStoreContext';
 import { SubscriptionSuggestionContext } from '../../../../Profile/Subscription/SubscriptionSuggestionContext';
 import { type ExampleShortHeader } from '../../../../Utils/GDevelopServices/Example';
 import Add from '../../../../UI/CustomSvgIcons/Add';
-import Skeleton from '@material-ui/lab/Skeleton';
-import BackgroundText from '../../../../UI/BackgroundText';
-import Paper from '../../../../UI/Paper';
 import PlaceholderError from '../../../../UI/PlaceholderError';
 import AlertMessage from '../../../../UI/AlertMessage';
 import IconButton from '../../../../UI/IconButton';
@@ -41,31 +35,14 @@ import { type PrivateGameTemplateListingData } from '../../../../Utils/GDevelopS
 import { PrivateGameTemplateStoreContext } from '../../../../AssetStore/PrivateGameTemplates/PrivateGameTemplateStoreContext';
 import ChevronArrowRight from '../../../../UI/CustomSvgIcons/ChevronArrowRight';
 import Refresh from '../../../../UI/CustomSvgIcons/Refresh';
-import ProjectFileListItem from './ProjectFileListItem';
-import { type MenuItemTemplate } from '../../../../UI/Menu/Menu.flow';
-import {
-  getExampleAndTemplateTiles,
-  getLastModifiedInfoByProjectId,
-  getProjectLineHeight,
-  transformCloudProjectsIntoFileMetadataWithStorageProviderName,
-} from './utils';
+import { getExampleAndTemplateTiles } from '../CreateSection/utils';
 import ErrorBoundary from '../../../../UI/ErrorBoundary';
 import InfoBar from '../../../../UI/Messages/InfoBar';
 import GridList from '@material-ui/core/GridList';
 import type { WindowSizeType } from '../../../../UI/Responsive/ResponsiveWindowMeasurer';
-import useAlertDialog from '../../../../UI/Alert/useAlertDialog';
-import optionalRequire from '../../../../Utils/OptionalRequire';
-import { deleteCloudProject } from '../../../../Utils/GDevelopServices/Project';
-import { extractGDevelopApiErrorStatusAndCode } from '../../../../Utils/GDevelopServices/Errors';
-import ContextMenu, {
-  type ContextMenuInterface,
-} from '../../../../UI/Menu/ContextMenu';
-import type { ClientCoordinates } from '../../../../Utils/UseLongTouch';
 import PromotionsSlideshow from '../../../../Promotions/PromotionsSlideshow';
 import ExampleStore from '../../../../AssetStore/ExampleStore';
-
-const electron = optionalRequire('electron');
-const path = optionalRequire('path');
+import ProjectFileList from '../CreateSection/ProjectFileList';
 
 const cellSpacing = 2;
 
@@ -120,13 +97,6 @@ type Props = {|
   closeProject: () => Promise<void>,
 |};
 
-const locateProjectFile = (file: FileMetadataAndStorageProviderName) => {
-  if (!electron) return;
-  electron.shell.showItemInFolder(
-    path.resolve(file.fileMetadata.fileIdentifier)
-  );
-};
-
 const BuildSection = ({
   project,
   currentFileMetadata,
@@ -142,15 +112,7 @@ const BuildSection = ({
   canManageGame,
   closeProject,
 }: Props) => {
-  const { getRecentProjectFiles } = React.useContext(PreferencesContext);
   const { exampleShortHeaders } = React.useContext(ExampleStoreContext);
-  const {
-    showDeleteConfirmation,
-    showConfirmation,
-    showAlert,
-  } = useAlertDialog();
-  const { removeRecentProjectFile } = React.useContext(PreferencesContext);
-  const [pendingProject, setPendingProject] = React.useState<?string>(null);
   const [
     showAllGameTemplates,
     setShowAllGameTemplates,
@@ -158,7 +120,6 @@ const BuildSection = ({
   const { privateGameTemplateListingDatas } = React.useContext(
     PrivateGameTemplateStoreContext
   );
-  const contextMenu = React.useRef<?ContextMenuInterface>(null);
   const authenticatedUser = React.useContext(AuthenticatedUserContext);
   const { openSubscriptionDialog } = React.useContext(
     SubscriptionSuggestionContext
@@ -170,170 +131,17 @@ const BuildSection = ({
   ] = React.useState<boolean>(false);
   const {
     authenticated,
-    profile,
-    cloudProjects,
     limits,
     cloudProjectsFetchingErrorLabel,
     onCloudProjectsChanged,
     onOpenLoginDialog,
   } = authenticatedUser;
   const { windowSize, isMobile, isLandscape } = useResponsiveWindowSize();
-  const [
-    lastModifiedInfoByProjectId,
-    setLastModifiedInfoByProjectId,
-  ] = React.useState({});
 
   const columnsCount = getItemsColumns(windowSize, isLandscape);
 
-  let projectFiles: Array<FileMetadataAndStorageProviderName> = getRecentProjectFiles().filter(
-    file => file.fileMetadata
-  );
-
-  if (cloudProjects) {
-    projectFiles = projectFiles.concat(
-      transformCloudProjectsIntoFileMetadataWithStorageProviderName(
-        cloudProjects
-      )
-    );
-  }
-
-  // Look at projects where lastCommittedBy is not the current user, and fetch
-  // public profiles of the users that have modified them.
-  React.useEffect(
-    () => {
-      const updateModificationInfoByProjectId = async () => {
-        if (!cloudProjects || !profile) return;
-
-        const _lastModifiedInfoByProjectId = await getLastModifiedInfoByProjectId(
-          {
-            cloudProjects,
-            profile,
-          }
-        );
-        setLastModifiedInfoByProjectId(_lastModifiedInfoByProjectId);
-      };
-
-      updateModificationInfoByProjectId();
-    },
-    [cloudProjects, profile]
-  );
-
   const hasTooManyCloudProjects = checkIfHasTooManyCloudProjects(
     authenticatedUser
-  );
-
-  const onDeleteCloudProject = async (
-    i18n: I18nType,
-    { fileMetadata, storageProviderName }: FileMetadataAndStorageProviderName
-  ) => {
-    if (storageProviderName !== 'Cloud') return;
-    const projectName = fileMetadata.name;
-    if (!projectName) return; // Only cloud projects can be deleted, and all cloud projects have names.
-
-    const isCurrentProjectOpened =
-      !!currentFileMetadata &&
-      currentFileMetadata.fileIdentifier === fileMetadata.fileIdentifier;
-    if (isCurrentProjectOpened) {
-      const result = await showConfirmation({
-        title: t`Project is opened`,
-        message: t`You are about to delete the project ${projectName}, which is currently opened. If you proceed, the project will be closed and you will lose any unsaved changes. Do you want to proceed?`,
-        confirmButtonLabel: t`Continue`,
-      });
-      if (!result) return;
-      await closeProject();
-    }
-
-    // Extract word translation to ensure it is not wrongly translated in the sentence.
-    const translatedConfirmText = i18n._(t`delete`);
-
-    const deleteAnswer = await showDeleteConfirmation({
-      title: t`Permanently delete the project?`,
-      message: t`Project ${projectName} will be deleted. You will no longer be able to access it.`,
-      fieldMessage: t`To confirm, type "${translatedConfirmText}"`,
-      confirmText: translatedConfirmText,
-      confirmButtonLabel: t`Delete project`,
-    });
-    if (!deleteAnswer) return;
-
-    try {
-      setPendingProject(fileMetadata.fileIdentifier);
-      await deleteCloudProject(authenticatedUser, fileMetadata.fileIdentifier);
-      authenticatedUser.onCloudProjectsChanged();
-    } catch (error) {
-      const extractedStatusAndCode = extractGDevelopApiErrorStatusAndCode(
-        error
-      );
-      const message =
-        extractedStatusAndCode && extractedStatusAndCode.status === 403
-          ? t`You don't have permissions to delete this project.`
-          : t`An error occurred when saving the project. Please try again later.`;
-      showAlert({
-        title: t`Unable to delete the project`,
-        message,
-      });
-    } finally {
-      setPendingProject(null);
-    }
-  };
-
-  const buildContextMenu = (
-    i18n: I18nType,
-    file: ?FileMetadataAndStorageProviderName
-  ): Array<MenuItemTemplate> => {
-    if (!file) return [];
-
-    const actions = [
-      { label: i18n._(t`Open`), click: () => onOpenRecentFile(file) },
-    ];
-    if (file.storageProviderName === 'Cloud') {
-      actions.push({
-        label: i18n._(t`Delete`),
-        click: () => onDeleteCloudProject(i18n, file),
-      });
-    } else if (file.storageProviderName === 'LocalFile') {
-      actions.push(
-        ...[
-          {
-            label: i18n._(t`Show in local folder`),
-            click: () => locateProjectFile(file),
-          },
-          {
-            label: i18n._(t`Remove from list`),
-            click: () => removeRecentProjectFile(file),
-          },
-        ]
-      );
-    } else {
-      actions.push({
-        label: i18n._(t`Remove from list`),
-        click: () => removeRecentProjectFile(file),
-      });
-    }
-
-    const gameId = file.fileMetadata.gameId;
-    if (gameId) {
-      actions.push(
-        ...[
-          { type: 'separator' },
-          {
-            label: i18n._(t`Manage game`),
-            click: () => onManageGame(gameId),
-            enabled: canManageGame(gameId),
-          },
-        ]
-      );
-    }
-
-    return actions;
-  };
-
-  const openContextMenu = React.useCallback(
-    (event: ClientCoordinates, file: FileMetadataAndStorageProviderName) => {
-      if (contextMenu.current) {
-        contextMenu.current.open(event.clientX, event.clientY, { file });
-      }
-    },
-    []
   );
 
   const refreshCloudProjects = React.useCallback(
@@ -353,12 +161,6 @@ const BuildSection = ({
     },
     [onCloudProjectsChanged, isRefreshing, authenticated]
   );
-
-  projectFiles.sort((a, b) => {
-    if (!a.fileMetadata.lastModifiedDate) return 1;
-    if (!b.fileMetadata.lastModifiedDate) return -1;
-    return b.fileMetadata.lastModifiedDate - a.fileMetadata.lastModifiedDate;
-  });
 
   const shouldDisplayPremiumGameTemplates =
     !authenticatedUser ||
@@ -398,7 +200,6 @@ const BuildSection = ({
     !authenticatedUser.limits.capabilities.classrooms ||
     !authenticatedUser.limits.capabilities.classrooms.hidePlayTab;
 
-  const skeletonLineHeight = getProjectLineHeight({ isMobile });
   const pageContent = showAllGameTemplates ? (
     <SectionContainer
       backAction={() => setShowAllGameTemplates(false)}
@@ -531,102 +332,15 @@ const BuildSection = ({
             </PlaceholderError>
           </Line>
         )}
-        {authenticatedUser.loginState === 'loggingIn' &&
-        projectFiles.length === 0 ? ( // Only show skeleton on first load
-          new Array(10).fill(0).map((_, index) => (
-            <ListItem style={styles.listItem} key={`skeleton-${index}`}>
-              <Line expand>
-                <Column expand>
-                  <Skeleton
-                    variant="rect"
-                    height={skeletonLineHeight}
-                    style={styles.projectSkeleton}
-                  />
-                </Column>
-              </Line>
-            </ListItem>
-          ))
-        ) : projectFiles.length > 0 ? (
-          <Line>
-            <Column noMargin expand>
-              {!isMobile && (
-                <Line justifyContent="space-between">
-                  <Column expand>
-                    <Text color="secondary">
-                      <Trans>File name</Trans>
-                    </Text>
-                  </Column>
-                  <Column expand>
-                    <Text color="secondary">
-                      <Trans>Location</Trans>
-                    </Text>
-                  </Column>
-                  <Column expand>
-                    <Text color="secondary">
-                      <Trans>Last edited</Trans>
-                    </Text>
-                  </Column>
-                </Line>
-              )}
-              <List>
-                {projectFiles.map(file => (
-                  <ProjectFileListItem
-                    key={file.fileMetadata.fileIdentifier}
-                    file={file}
-                    onOpenContextMenu={openContextMenu}
-                    isLoading={
-                      pendingProject === file.fileMetadata.fileIdentifier
-                    }
-                    currentFileMetadata={currentFileMetadata}
-                    storageProviders={storageProviders}
-                    isWindowSizeMediumOrLarger={!isMobile}
-                    onOpenRecentFile={onOpenRecentFile}
-                    lastModifiedInfo={
-                      lastModifiedInfoByProjectId[
-                        file.fileMetadata.fileIdentifier
-                      ]
-                    }
-                  />
-                ))}
-                {isMobile && limits && hasTooManyCloudProjects && (
-                  <MaxProjectCountAlertMessage
-                    margin="dense"
-                    limits={limits}
-                    onUpgrade={() =>
-                      openSubscriptionDialog({
-                        analyticsMetadata: {
-                          reason: 'Cloud Project limit reached',
-                        },
-                      })
-                    }
-                  />
-                )}
-              </List>
-            </Column>
-          </Line>
-        ) : (
-          <Line>
-            <ListItem style={styles.listItem}>
-              <Column expand>
-                <Paper
-                  variant="outlined"
-                  background="dark"
-                  style={styles.noProjectsContainer}
-                >
-                  <BackgroundText>
-                    <Trans>No projects yet.</Trans>
-                  </BackgroundText>
-                  <BackgroundText>
-                    <Trans>
-                      Create your first project using one of our templates or
-                      start from scratch.
-                    </Trans>
-                  </BackgroundText>
-                </Paper>
-              </Column>
-            </ListItem>
-          </Line>
-        )}
+        <ProjectFileList
+          i18n={i18n}
+          game={null}
+          onOpenProject={onOpenRecentFile}
+          storageProviders={storageProviders}
+          project={project}
+          currentFileMetadata={currentFileMetadata}
+          closeProject={closeProject}
+        />
       </SectionRow>
       <SectionRow>
         <Line alignItems="center" noMargin expand>
@@ -658,10 +372,6 @@ const BuildSection = ({
         duration={5000}
         onActionClick={onOpenLoginDialog}
         actionLabel={<Trans>Log in</Trans>}
-      />
-      <ContextMenu
-        ref={contextMenu}
-        buildMenuTemplate={(_i18n, { file }) => buildContextMenu(_i18n, file)}
       />
     </>
   );
