+++ newIDE/app/src/MainFrame/EditorContainers/HomePage/index.js
@@ -463,6 +463,10 @@ export const HomePage = React.memo<Props>(
                   {activeTab === 'manage' && (
                     <ManageSection
                       project={project}
+                      currentFileMetadata={fileMetadata}
+                      onOpenProject={onOpenRecentFile}
+                      storageProviders={storageProviders}
+                      closeProject={closeProject}
                       games={games}
                       onRefreshGames={fetchGames}
                       onGameUpdated={onGameUpdated}
