+++ newIDE/app/src/ProjectsStorage/CloudStorageProvider/CloudProjectWriter.js
@@ -215,6 +215,7 @@ export const generateOnChooseSaveProjectAsLocation = ({
   return {
     saveAsLocation: {
       name,
+      gameId: project.getProjectUuid(),
     },
   };
 };
