+++ newIDE/app/src/stories/componentStories/GameDashboard/GameDashboard.stories.js
@@ -8,7 +8,7 @@ import { action } from '@storybook/addon-actions';
 import paperDecorator from '../../PaperDecorator';
 
 import GameDashboard from '../../../GameDashboard';
-
+import CloudStorageProvider from '../../../ProjectsStorage/CloudStorageProvider';
 import { type AuthenticatedUser } from '../../../Profile/AuthenticatedUserContext';
 import {
   game1,
@@ -324,8 +324,10 @@ export const Default = ({
     <AuthenticatedUserContext.Provider value={authenticatedUser}>
       <MarketingPlansStoreStateProvider>
         <GameDashboard
+          currentFileMetadata={null}
+          onOpenProject={action('onOpenProject')}
+          storageProviders={[CloudStorageProvider]}
           game={game}
-          analyticsSource="homepage"
           key={renderCount.toFixed(0)}
           currentView={tab}
           setCurrentView={setTab}
@@ -339,6 +341,7 @@ export const Default = ({
             )
           }
           gameUnregisterErrorText={gameUnregisterErrorText}
+          closeProject={action('closeProject')}
         />
       </MarketingPlansStoreStateProvider>
     </AuthenticatedUserContext.Provider>
