+++ newIDE/app/src/stories/componentStories/GameDashboard/GamesList.stories.js
@@ -12,6 +12,7 @@ import {
 } from '../../../fixtures/GDevelopServicesTestData';
 import AuthenticatedUserContext from '../../../Profile/AuthenticatedUserContext';
 import GamesList from '../../../GameDashboard/GamesList';
+import CloudStorageProvider from '../../../ProjectsStorage/CloudStorageProvider';
 
 export default {
   title: 'GameDashboard/GamesList',
@@ -23,10 +24,12 @@ export const WithoutAProjectOpened = () => {
   return (
     <AuthenticatedUserContext.Provider value={fakeSilverAuthenticatedUser}>
       <GamesList
+        storageProviders={[CloudStorageProvider]}
         project={null}
         games={[game1, game2]}
         onRefreshGames={action('onRefreshGames')}
         onOpenGameId={action('onOpenGameId')}
+        onOpenProject={action('onOpenProject')}
       />
     </AuthenticatedUserContext.Provider>
   );
